feat: Add icon support for operations
